Add authorization apis - JWT token validate added - Fixed some bugs

+++ controllers/User.php
@@ -3,6 +3,7 @@
 include_once $_SERVER['DOCUMENT_ROOT'] . '/utils/jwt.php';
 include_once $_SERVER['DOCUMENT_ROOT'] . '/utils/File_Manager.php';
 include_once $_SERVER['DOCUMENT_ROOT'] . '/utils/utils.php';
+include_once $_SERVER['DOCUMENT_ROOT'] . '/utils/mail.php';
 
 class User
 {
@@ -21,17 +22,14 @@ class User
 
     public function GetUser($auth, $username){
         try {
-            //todo: auth token is important
-//            if($this->jwt->Validate_Token($auth)) {
-
-            $get_result = $this->GetUserByUsername($username);
-
-            if ($get_result != null && count($get_result) > 0) {
-                return $get_result;
-            } else {
-                return 'not_found';
+            if($this->jwt->Validate_Token($auth)) {
+                $get_result = $this->GetUserByUsername($username);
+                if ($get_result != null && count($get_result) > 0) {
+                    return $get_result;
+                } else {
+                    return 'not_found';
+                }
             }
-//            }
         } catch (PDOException $pdo_exception) {
             return 'PDO : ' . $pdo_exception->getMessage();
         } catch (Exception $exception) {
@@ -41,8 +39,7 @@ class User
 
     public function GetAllUser($auth){
         try {
-            //todo: validate token
-//            if($this->jwt->Validate_Token($auth)) {
+            if($this->jwt->Validate_Token($auth)) {
                 $query = sprintf("SELECT * FROM %s ORDER BY id", self::$table_name);
 
                 $this->conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
@@ -57,15 +54,15 @@ class User
                 } else {
                     return false;
                 }
-//            }else{
-//                return false;
-//            }
+            }else{
+                return false;
+            }
         } catch (PDOException $pdo_exception){
             return 'PDO Exception : ' . $pdo_exception->getMessage();
         }
     }
 
-    private function GetUserById($id){
+    public function GetUserById($id){
         try {
             $query = sprintf("SELECT * FROM %s WHERE id = :id", self::$table_name);
 
@@ -89,7 +86,7 @@ class User
         }
     }
 
-    private function GetUserByEmail($email){
+    public function GetUserByEmail($email){
         try {
             $query = sprintf("SELECT * FROM %s WHERE email=:email", self::$table_name);
 
@@ -113,14 +110,12 @@ class User
         }
     }
 
-    private function GetUserByUsername($username){
+    public function GetUserByUsername($username){
         try {
             $query = "SELECT * FROM " . self::$table_name . " WHERE username LIKE '%" . $username . "%'";
 
             $this->conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
             $stmt = $this->conn->prepare($query);
-
-//            $stmt->bindParam(':username',$username);
 
             if ($stmt->execute()) {
                 if ($stmt->rowCount()) {
@@ -137,7 +132,7 @@ class User
         }
     }
 
-    private function GetUserByPhone($phone){
+    public function GetUserByPhone($phone){
         try {
             $query = sprintf("SELECT * FROM %s WHERE phone=:phone", self::$table_name);
 
@@ -164,8 +159,7 @@ class User
     public function AddUser($auth, $args)
     {
         try {
-            //todo: auth token is important
-//            if($this->jwt->Validate_Token($auth)) {
+            if($this->jwt->Validate_Token($auth)) {
                 $username = $args['username'];
                 $password = md5($args['password']);
                 $full_name = empty($args['full_name']) ? '' : $args['full_name'];
@@ -182,7 +176,7 @@ class User
                 if(!empty($args['user_avatar']))
                     $user_avatar = $this->file_manager->Upload_Image($user_avatar);
 
-                $query = sprintf("INSERT INTO %s SET username=:username, password=:password, full_name=:full_name, email=:email, user_avatar=:user_avatar, phone=:phone, active_token=:active_token, is_active=0", self::$table_name);
+                $query = sprintf("INSERT INTO %s (username,password,full_name,email,user_avatar,phone,active_token,is_active) VALUES (:username,:password,:full_name,:email,:user_avatar,:phone,:active_token,0)", self::$table_name);
 
                 $this->conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
                 $stmt = $this->conn->prepare($query);
@@ -195,18 +189,20 @@ class User
                 $stmt->bindParam(':phone', $phone);
                 $stmt->bindParam(':active_token', $active_token);
 
-                //todo: send verification email
-
                 if ($stmt->execute()) {
                     if ($stmt->rowCount()) {
-                        return $this->GetUserById($this->conn->lastInsertId());
+                        $user = $this->GetUserById($this->conn->lastInsertId());
+                        send_mail($user['email'], 'فعالسازی حساب کاربری', 'حساب کاربری شما غیرفعال میباشد. برای فعالسازی لطفاً کد زیر را در اپلیکشین وارد کنید : ' . $user['active_token']);
+                        return $user;
                     } else {
                         return 'failed_user_add';
                     }
                 } else {
                     return false;
                 }
-//            }
+            }else{
+                return false;
+            }
         } catch (PDOException $pdo_exception) {
             return 'PDO : ' . $pdo_exception->getMessage();
         } catch (Exception $exception) {
@@ -216,51 +212,52 @@ class User
 
     public function EditUser($auth, $userId, $args = array()){
         try {
-            //todo: auth token is important
-//            if($this->jwt->Validate_Token($auth)) {
+            if($this->jwt->Validate_Token($auth)) {
 
-            $user_old_data = $this->GetUserById($userId);
-            if ($user_old_data == null)
-                return 'user_not_found';
+                $user_old_data = $this->GetUserById($userId);
+                if ($user_old_data == null)
+                    return 'user_not_found';
 
-            $username = isset($args['username']) && !empty($args['username']) ? $args['username'] : $user_old_data['username'];
-            $password = isset($args['password']) && !empty($args['password']) ? md5($args['password']) : $user_old_data['password'];
-            $full_name = isset($args['full_name']) && !empty($args['full_name']) ? $args['full_name'] : $user_old_data['full_name'];
-            $email = isset($args['email']) && !empty($args['email']) ? $args['email'] : $user_old_data['email'];
-            $user_avatar = isset($args['user_avatar']) && !empty($args['user_avatar']) ? $args['user_avatar'] : $user_old_data['user_avatar'];
-            $phone = isset($args['phone']) && !empty($args['phone']) ? $args['phone'] : $user_old_data['phone'];
-            $active_token = isset($args['active_token']) && !empty($args['active_token']) ? $args['active_token'] : $user_old_data['active_token'];
+                $username = isset($args['username']) && !empty($args['username']) ? $args['username'] : $user_old_data['username'];
+                $password = isset($args['password']) && !empty($args['password']) ? md5($args['password']) : $user_old_data['password'];
+                $full_name = isset($args['full_name']) && !empty($args['full_name']) ? $args['full_name'] : $user_old_data['full_name'];
+                $email = isset($args['email']) && !empty($args['email']) ? $args['email'] : $user_old_data['email'];
+                $user_avatar = isset($args['user_avatar']) && !empty($args['user_avatar']) ? $args['user_avatar'] : $user_old_data['user_avatar'];
+                $phone = isset($args['phone']) && !empty($args['phone']) ? $args['phone'] : $user_old_data['phone'];
+                $active_token = isset($args['active_token']) && !empty($args['active_token']) ? $args['active_token'] : $user_old_data['active_token'];
 //            $is_active = isset($args['is_active']) && !empty($args['is_active']) ? $args['is_active'] : $user_old_data['is_active'];
 
-            if(isset($args['user_avatar']) && !empty($args['user_avatar'])) {
-                $this->file_manager->Remove_Old_Image($user_old_data['user_avatar']);
-                $user_avatar = $this->file_manager->Upload_Image($user_avatar);
-            }
-
-            $query = sprintf("UPDATE %s SET username=:username, password=:password, full_name=:full_name, email=:email, user_avatar=:user_avatar, phone=:phone, active_token=:active_token WHERE id=:id", self::$table_name);
-
-            $this->conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-            $stmt = $this->conn->prepare($query);
-
-            $stmt->bindParam(':id', $userId);
-            $stmt->bindParam(':username', $username);
-            $stmt->bindParam(':password', $password);
-            $stmt->bindParam(':full_name', $full_name);
-            $stmt->bindParam(':email', $email);
-            $stmt->bindParam(':user_avatar', $user_avatar);
-            $stmt->bindParam(':phone', $phone);
-            $stmt->bindParam(':active_token', $active_token);
-
-            if ($stmt->execute()) {
-                if ($stmt->rowCount()) {
-                    return $this->GetUserById($userId);
-                } else {
-                    return 'failed_user_add';
+                if (isset($args['user_avatar']) && !empty($args['user_avatar'])) {
+                    $this->file_manager->Remove_Old_Image($user_old_data['user_avatar']);
+                    $user_avatar = $this->file_manager->Upload_Image($user_avatar);
                 }
-            } else {
-                return false;
+
+                $query = sprintf("UPDATE %s SET username=:username, password=:password, full_name=:full_name, email=:email, user_avatar=:user_avatar, phone=:phone, active_token=:active_token WHERE id=:id", self::$table_name);
+
+                $this->conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+                $stmt = $this->conn->prepare($query);
+
+                $stmt->bindParam(':id', $userId);
+                $stmt->bindParam(':username', $username);
+                $stmt->bindParam(':password', $password);
+                $stmt->bindParam(':full_name', $full_name);
+                $stmt->bindParam(':email', $email);
+                $stmt->bindParam(':user_avatar', $user_avatar);
+                $stmt->bindParam(':phone', $phone);
+                $stmt->bindParam(':active_token', $active_token);
+
+                if ($stmt->execute()) {
+                    if ($stmt->rowCount()) {
+                        return $this->GetUserById($userId);
+                    } else {
+                        return 'failed_user_add';
+                    }
+                } else {
+                    return false;
+                }
+            }else{
+                return  false;
             }
-//            }
         } catch (PDOException $pdo_exception) {
             return 'PDO : ' . $pdo_exception->getMessage();
         } catch (Exception $exception) {
@@ -270,8 +267,7 @@ class User
 
     public function DeleteUser($auth, $userId){
         try {
-            //todo : validation token is here
-//            if($this->jwt->Validate_Token($auth)){
+            if($this->jwt->Validate_Token($auth)){
                 $get_result = $this->GetUserById($userId);
 
                 if ($get_result != null && !empty($get_result)) {
@@ -290,9 +286,9 @@ class User
                 } else {
                     return 'user_not_found';
                 }
-//            }else{
-//                return false;
-//            }
+            }else{
+                return false;
+            }
         } catch (PDOException $pdo_exception) {
             return 'PDO : ' . $pdo_exception->getMessage();
         } catch (Exception $exception) {
@@ -302,59 +298,60 @@ class User
 
     public function FollowUnfollowUser($auth, $userId, $followerId){
         try {
-            //todo: auth token is important
-//            if($this->jwt->Validate_Token($auth)) {
+            if($this->jwt->Validate_Token($auth)) {
 
-            if($userId == $followerId)
-                return 'user_same_as_follower';
+                if ($userId == $followerId)
+                    return 'user_same_as_follower';
 
-            $get_user_result = $this->GetUserById($userId);
-            if ($get_user_result == null)
-                return 'user_not_found';
+                $get_user_result = $this->GetUserById($userId);
+                if ($get_user_result == null)
+                    return 'user_not_found';
 
-            $get_follower_result = $this->GetUserById($followerId);
-            if ($get_follower_result == null)
-                return 'follower_not_found';
+                $get_follower_result = $this->GetUserById($followerId);
+                if ($get_follower_result == null)
+                    return 'follower_not_found';
 
-            $get_follower = $this->GetFollower($userId,$followerId);
-            if($get_follower != null){
-                $query = sprintf("DELETE FROM %s WHERE user_id=:user_id AND follower_id=:follower_id", self::$Follower_table_name);
+                $get_follower = $this->GetFollower($userId, $followerId);
+                if ($get_follower != null) {
+                    $query = sprintf("DELETE FROM %s WHERE user_id=:user_id AND follower_id=:follower_id", self::$Follower_table_name);
 
-                $this->conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-                $stmt = $this->conn->prepare($query);
+                    $this->conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+                    $stmt = $this->conn->prepare($query);
 
-                $stmt->bindParam(':user_id', $userId);
-                $stmt->bindParam(':follower_id', $followerId);
+                    $stmt->bindParam(':user_id', $userId);
+                    $stmt->bindParam(':follower_id', $followerId);
 
-                if ($stmt->execute()) {
-                    if ($stmt->rowCount()) {
-                        return 'unfollow_successful';
+                    if ($stmt->execute()) {
+                        if ($stmt->rowCount()) {
+                            return 'unfollow_successful';
+                        } else {
+                            return 'failed_user_unfollow';
+                        }
                     } else {
-                        return 'failed_user_unfollow';
+                        return false;
                     }
                 } else {
-                    return false;
+                    $query = sprintf("INSERT INTO %s SET user_id=:user_id, follower_id=:follower_id", self::$Follower_table_name);
+
+                    $this->conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+                    $stmt = $this->conn->prepare($query);
+
+                    $stmt->bindParam(':user_id', $userId);
+                    $stmt->bindParam(':follower_id', $followerId);
+
+                    if ($stmt->execute()) {
+                        if ($stmt->rowCount()) {
+                            return 'follow_successful';
+                        } else {
+                            return 'failed_user_follow';
+                        }
+                    } else {
+                        return false;
+                    }
                 }
             }else{
-                $query = sprintf("INSERT INTO %s SET user_id=:user_id, follower_id=:follower_id", self::$Follower_table_name);
-
-                $this->conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-                $stmt = $this->conn->prepare($query);
-
-                $stmt->bindParam(':user_id', $userId);
-                $stmt->bindParam(':follower_id', $followerId);
-
-                if ($stmt->execute()) {
-                    if ($stmt->rowCount()) {
-                        return 'follow_successful';
-                    } else {
-                        return 'failed_user_follow';
-                    }
-                } else {
-                    return false;
-                }
+                return false;
             }
-//            }
         } catch (PDOException $pdo_exception) {
             return 'PDO : ' . $pdo_exception->getMessage();
         } catch (Exception $exception) {
